requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.13
pydantic==2.6.1
PyMuPDF==1.23.25
python-dotenv==1.0.1
chromadb==0.4.24
elevenlabs==0.2.27
openai==1.14.2
python-multipart==0.0.9
sentence_transformers==2.6.0